+++ CHANGES.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Clamp components of `Color` class automatically
 - `utils.computeVertexNormal` does not return inverted normals anymore
+- Allow negative values for `Shape:setBorderWidth` to create inner borders
 
 ### Fixed